requirements.txt
cycler==0.10.0
joblib==0.13.2
jsons==0.8.8
kiwisolver==1.1.0
matplotlib==3.0.3
numpy==1.16.3
pandas==0.24.2
pprofile==2.0.2
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
scipy==1.2.1
seaborn==0.9.0
six==1.12.0
tqdm==4.36.1
psutil